refactor(exampleUtils): extract formatToRelativeString from date utils
